+++ src/modules/modalities/entities/Modalities.ts
@@ -1,0 +1,3 @@
+class Modalities {}
+
+export { Modalities };
