+++ src/shared/container/index.ts
@@ -1,10 +1,10 @@
 import "reflect-metadata";
 import { container } from "tsyringe";
 
-import { ICourtsRepository } from "../../modules/repositories/ICourtsRepository";
-import { CourtsRepository } from "../../modules/repositories/implementation/CourtsRespository";
-import { UsersRepository } from "../../modules/repositories/implementation/UsersRepository";
-import { IUserRepository } from "../../modules/repositories/IUserRepository";
+import { UsersRepository } from "../../modules/accounts/repositories/implementation/UsersRepository";
+import { IUserRepository } from "../../modules/accounts/repositories/IUserRepository";
+import { ICourtsRepository } from "../../modules/courts/repositories/ICourtsRepository";
+import { CourtsRepository } from "../../modules/courts/repositories/implementation/CourtsRespository";
 
 container.registerSingleton<ICourtsRepository>(
   "CourtsRepository",
